+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.4.13] - 2026-03-26
+
+### Documentation
+- **Info Command Help**: Added help text for `agenthub info` command
+- **Publish-Remote Command Help**: Added help text for `agenthub publish-remote` command
+- All 19 CLI commands now have complete help documentation
+
 ## [0.4.12] - 2026-03-26
 
 ### Added